feat: resolutions WIP
